// src/types/supabase.ts
export type Database = {
// ...
  public: {
    Tables: {
      bookings: {
        Row: {
          cabinID: number
          cabinPrice: number
          created_at: string
          endDate: string
          extrasPrice: number
          guestID: number | null
          hasBreakfast: boolean
          id: number
          isPaid: boolean
          numGuests: number
          numNights: number
          observations: string
          startDate: string
          status: string
          totalPrice: number
        }
        Insert: {
          cabinID: number
          cabinPrice: number
          created_at?: string
          endDate: string
          extrasPrice: number
          guestID?: number | null
          hasBreakfast: boolean
          id?: number
          isPaid: boolean
          numGuests: number
          numNights: number
          observations: string
          startDate: string
          status: string
          totalPrice: number
        }
        Update: {
          cabinID?: number
          cabinPrice?: number
          created_at?: string
          endDate?: string
          extrasPrice?: number
          guestID?: number | null
          hasBreakfast?: boolean
          id?: number
          isPaid?: boolean
          numGuests?: number
          numNights?: number
          observations?: string
          startDate?: string
          status?: string
          totalPrice?: number
        }
        Relationships: [
          {
            foreignKeyName: "bookings_cabinID_fkey"
            columns: ["cabinID"]
            isOneToOne: false
            referencedRelation: "cabins"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "bookings_guestID_fkey"
            columns: ["guestID"]
            isOneToOne: false
            referencedRelation: "guests"
            referencedColumns: ["id"]
          },
        ]
      }
      cabins: {
        Row: {
          created_at: string
          description: string
          discount: number
          id: number
          image: string
          maxCapacity: number
          name: string
          regularPrice: number
        }
        Insert: {
          created_at?: string
          description: string
          discount: number
          id?: number
          image: string
          maxCapacity: number
          name: string
          regularPrice: number
        }
        Update: {
          created_at?: string
          description?: string
          discount?: number
          id?: number
          image?: string
          maxCapacity?: number
          name?: string
          regularPrice?: number
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never